Move MSF API App and associated servlets
The modules interact with the DbManager, however, are not a part of it
and belong in a more meaningful location for web services.
