--- lib/msf/core/db_manager/http/authentication.rb
@@ -1,5 +1,0 @@
-module Authentication
-  autoload :Strategies, 'msf/core/db_manager/http/authentication/strategies'
-
-  include Strategies
-end
