+++ lib/msf/core/web_services/authentication.rb
@@ -1,0 +1,5 @@
+module Authentication
+  autoload :Strategies, 'msf/core/web_services/authentication/strategies'
+
+  include Strategies
+end
