+++ lib/msf/core/web_services/authentication/strategies.rb
@@ -1,8 +1,8 @@
 module Authentication
   module Strategies
-    autoload :ApiToken, 'msf/core/db_manager/http/authentication/strategies/api_token'
-    autoload :AdminApiToken, 'msf/core/db_manager/http/authentication/strategies/admin_api_token'
-    autoload :UserPassword, 'msf/core/db_manager/http/authentication/strategies/user_password'
+    autoload :ApiToken, 'msf/core/web_services/authentication/strategies/api_token'
+    autoload :AdminApiToken, 'msf/core/web_services/authentication/strategies/admin_api_token'
+    autoload :UserPassword, 'msf/core/web_services/authentication/strategies/user_password'
 
     Warden::Strategies.add(:api_token, Authentication::Strategies::ApiToken)
     Warden::Strategies.add(:admin_api_token, Authentication::Strategies::AdminApiToken)
