+++ lib/msf/core/web_services/authentication/strategies/admin_api_token.rb
