+++ lib/msf/core/web_services/authentication/strategies/api_token.rb
