+++ lib/msf/core/web_services/authentication/strategies/user_password.rb
