+++ lib/msf/core/web_services/db_manager_proxy.rb
