+++ lib/msf/core/web_services/job_processor.rb
