+++ lib/msf/core/web_services/metasploit_api_app.rb
@@ -2,28 +2,28 @@ require 'sinatra/base'
 require 'swagger/blocks'
 require 'sysrandom/securerandom'
 require 'warden'
-require 'msf/core/db_manager/http/authentication'
-require 'msf/core/db_manager/http/servlet_helper'
-require 'msf/core/db_manager/http/servlet/api_docs_servlet'
-require 'msf/core/db_manager/http/servlet/auth_servlet'
-require 'msf/core/db_manager/http/servlet/host_servlet'
-require 'msf/core/db_manager/http/servlet/note_servlet'
-require 'msf/core/db_manager/http/servlet/vuln_servlet'
-require 'msf/core/db_manager/http/servlet/event_servlet'
-require 'msf/core/db_manager/http/servlet/web_servlet'
-require 'msf/core/db_manager/http/servlet/msf_servlet'
-require 'msf/core/db_manager/http/servlet/workspace_servlet'
-require 'msf/core/db_manager/http/servlet/service_servlet'
-require 'msf/core/db_manager/http/servlet/session_servlet'
-require 'msf/core/db_manager/http/servlet/exploit_servlet'
-require 'msf/core/db_manager/http/servlet/loot_servlet'
-require 'msf/core/db_manager/http/servlet/session_event_servlet'
-require 'msf/core/db_manager/http/servlet/credential_servlet'
-require 'msf/core/db_manager/http/servlet/login_servlet'
-require 'msf/core/db_manager/http/servlet/nmap_servlet'
-require 'msf/core/db_manager/http/servlet/db_export_servlet'
-require 'msf/core/db_manager/http/servlet/vuln_attempt_servlet'
-require 'msf/core/db_manager/http/servlet/user_servlet'
+require 'msf/core/web_services/authentication'
+require 'msf/core/web_services/servlet_helper'
+require 'msf/core/web_services/servlet/api_docs_servlet'
+require 'msf/core/web_services/servlet/auth_servlet'
+require 'msf/core/web_services/servlet/host_servlet'
+require 'msf/core/web_services/servlet/note_servlet'
+require 'msf/core/web_services/servlet/vuln_servlet'
+require 'msf/core/web_services/servlet/event_servlet'
+require 'msf/core/web_services/servlet/web_servlet'
+require 'msf/core/web_services/servlet/msf_servlet'
+require 'msf/core/web_services/servlet/workspace_servlet'
+require 'msf/core/web_services/servlet/service_servlet'
+require 'msf/core/web_services/servlet/session_servlet'
+require 'msf/core/web_services/servlet/exploit_servlet'
+require 'msf/core/web_services/servlet/loot_servlet'
+require 'msf/core/web_services/servlet/session_event_servlet'
+require 'msf/core/web_services/servlet/credential_servlet'
+require 'msf/core/web_services/servlet/login_servlet'
+require 'msf/core/web_services/servlet/nmap_servlet'
+require 'msf/core/web_services/servlet/db_export_servlet'
+require 'msf/core/web_services/servlet/vuln_attempt_servlet'
+require 'msf/core/web_services/servlet/user_servlet'
 require 'msf/core/web_services/servlet/module_search_servlet'
 
 class MetasploitApiApp < Sinatra::Base
