+++ lib/msf/core/web_services/public/favicon-16x16.png
