+++ lib/msf/core/web_services/public/favicon-32x32.png
