+++ lib/msf/core/web_services/public/favicon.ico
