+++ lib/msf/core/web_services/public/msf-ws.css
