+++ lib/msf/core/web_services/public/oauth2-redirect.html
