+++ lib/msf/core/web_services/public/swagger-ui.css
