+++ lib/msf/core/web_services/servlet/api_docs_servlet.rb
