+++ lib/msf/core/web_services/servlet/auth_servlet.rb
