+++ lib/msf/core/web_services/servlet/credential_servlet.rb
