+++ lib/msf/core/web_services/servlet/event_servlet.rb
