+++ lib/msf/core/web_services/servlet/host_servlet.rb
