+++ lib/msf/core/web_services/servlet/login_servlet.rb
