+++ lib/msf/core/web_services/servlet/loot_servlet.rb
