+++ lib/msf/core/web_services/servlet/msf_servlet.rb
