+++ lib/msf/core/web_services/servlet/note_servlet.rb
