+++ lib/msf/core/web_services/servlet/service_servlet.rb
