+++ lib/msf/core/web_services/servlet/session_event_servlet.rb
