+++ lib/msf/core/web_services/servlet/session_servlet.rb
