+++ lib/msf/core/web_services/servlet/user_servlet.rb
