+++ lib/msf/core/web_services/servlet/vuln_attempt_servlet.rb
