+++ lib/msf/core/web_services/servlet/vuln_servlet.rb
