+++ lib/msf/core/web_services/servlet/web_servlet.rb
