+++ lib/msf/core/web_services/servlet/workspace_servlet.rb
