+++ lib/msf/core/web_services/views/api_docs.erb
