+++ lib/msf/core/web_services/views/auth/account.erb
