+++ lib/msf/core/web_services/views/auth/login.erb
