+++ msfdb
@@ -422,7 +422,7 @@ def init_web_service
 
       require File.expand_path('./config/boot', @framework_path)
       require 'metasploit/framework/parsed_options/remote_db'
-      require 'msf/core/db_manager/http/metasploit_api_app'
+      require 'msf/core/web_services/metasploit_api_app'
 
       def require_environment!(parsed_options)
         # RAILS_ENV must be set before requiring 'config/application.rb'
