+++ msfdb_ws
@@ -16,7 +16,7 @@ end
 def require_deps
   require 'pathname'
   require Pathname.new(__FILE__).realpath.expand_path.parent.join('config', 'boot')
-  require 'msf/core/db_manager/http/http_db_manager_service'
+  require 'msf/core/web_services/http_db_manager_service'
 end
 
 def parse_args(args)
